moved requestAudio from Hearing Service to Audio Request Service
