+++ src/app/portal/services/audio-request/audio-request.service.spec.ts
@@ -1,6 +1,6 @@
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { TestBed, discardPeriodicTasks, fakeAsync, tick } from '@angular/core/testing';
-import { RequestedMedia, RequestedMediaData } from '@portal-types/index';
+import { PostAudioRequest, RequestedMedia, RequestedMediaData } from '@portal-types/index';
 import { DateTime, Settings } from 'luxon';
 import { of } from 'rxjs';
 import { AudioRequestService } from './audio-request.service';
@@ -266,6 +266,66 @@ describe('AudioService', () => {
     });
 
     expect(receivedBlob).toBeInstanceOf(Blob);
+  });
+
+  describe('#requestAudio', () => {
+    it('should request audio for download', () => {
+      let response;
+      const audioRequest: PostAudioRequest = {
+        hearing_id: 1,
+        requestor: 1,
+        start_time: '2023-09-01T02:00:00Z',
+        end_time: '2023-09-01T15:32:24Z',
+        request_type: 'DOWNLOAD',
+      };
+      const mockResponse = {
+        request_id: 141,
+        case_id: 'DMP461_Case12',
+        courthouse_name: 'LIVERPOOL_DMP461',
+        hearing_date: '2023-08-11',
+        start_time: '2023-08-11T15:30:17Z',
+        end_time: '2023-08-11T15:30:17Z',
+      };
+
+      service.requestAudio(audioRequest).subscribe((res) => (response = res));
+
+      const req = httpMock.expectOne((request) => {
+        return request.body === audioRequest && request.url === 'api/audio-requests/download';
+      });
+      expect(req.request.method).toBe('POST');
+      req.flush(mockResponse);
+
+      expect(response).toEqual(mockResponse);
+    });
+
+    it('should request audio for playback', () => {
+      let response;
+      const audioRequest: PostAudioRequest = {
+        hearing_id: 1,
+        requestor: 1,
+        start_time: '2023-09-01T02:00:00Z',
+        end_time: '2023-09-01T15:32:24Z',
+        request_type: 'PLAYBACK',
+      };
+      const mockResponse = {
+        request_id: 141,
+        case_id: 'DMP461_Case12',
+        courthouse_name: 'LIVERPOOL_DMP461',
+        hearing_date: '2023-08-11',
+        start_time: '2023-08-11T15:30:17Z',
+        end_time: '2023-08-11T15:30:17Z',
+      };
+
+      service.requestAudio(audioRequest).subscribe((res) => (response = res));
+
+      const req = httpMock.expectOne((request) => {
+        return request.body === audioRequest && request.url === 'api/audio-requests/playback';
+      });
+      expect(req.request.method).toBe('POST');
+      req.flush(mockResponse);
+
+      expect(response).toEqual(mockResponse);
+    });
   });
 
   describe('#getStatusCode', () => {
